added new functions: getMostServedRoute_byAirport getMostServedAirline_byAirport

+++ src/main/java/com/task2_3/server/RankKey.java
@@ -1,15 +1,15 @@
 package com.task2_3.server;
 
 public class RankKey {
-    double airport_id;
+    int airport_id;
     String carrier_id;
 
-    public RankKey(double airport_id, String carrier_id){
+    public RankKey(int airport_id, String carrier_id){
         this.airport_id = airport_id;
         this.carrier_id = carrier_id;
     }
 
-    public double getAirport(){
+    public int getAirport(){
         return airport_id;
     }
